requirements.txt
PyQt5==5.15.9
speedtest-cli==2.1.3